Enhance LatencyMonitor to utilize Datadog DDSketch for quantile calculations; update requirements and modify producer to handle latency as an integer.

+++ src/metrics_streamer/producer.py
@@ -44,7 +44,7 @@ class LogProducer:
                 for _, row in chunk.iterrows():
                     data = {
                         'block_id': row['BlockId'],
-                        'latency': float(row['Latency'])
+                        'latency': int(row['Latency'])
                     }
                     self.send_metric(data)
                     logger.debug(f"Sent metric: {data}")
